update to .net 7 integration tests for azure db added implementation of iequatable renamed a bunch of things. fixed a few db related issues.

+++ FTMContextNet/Application/Models/Create/CreateImportModel.cs
@@ -10,7 +10,7 @@ namespace FTMContextNet.Application.Models.Create
     {
         public string FileName { get; set; }
 
-        public double FileSize { get; set; }
+        public long FileSize { get; set; }
 
         public bool Selected { get; set; }
